Refactor Config registry strategy
The mod is to own the configurator, not the config per se.
