+++ src/main/java/com/github/mkram17/bazaarutils/BazaarUtils.java
@@ -2,18 +2,21 @@ package com.github.mkram17.bazaarutils;
 
 import com.github.mkram17.bazaarutils.config.BUConfig;
 import com.github.mkram17.bazaarutils.config.util.ConfigUtil;
-import com.github.mkram17.bazaarutils.events.listener.BUListener;
 import com.github.mkram17.bazaarutils.events.listener.ListenerManager;
 import com.github.mkram17.bazaarutils.features.Bookmark;
 import com.github.mkram17.bazaarutils.misc.BUCompatibilityHelper;
 import com.github.mkram17.bazaarutils.utils.BUCommands;
 import com.mojang.serialization.Codec;
+import com.teamresourceful.resourcefulconfig.api.annotations.Config;
+import com.teamresourceful.resourcefulconfig.api.loader.Configurator;
+import com.teamresourceful.resourcefulconfig.api.types.ResourcefulConfig;
 import lombok.Getter;
 import meteordevelopment.orbit.EventBus;
 import meteordevelopment.orbit.IEventBus;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback;
 import net.fabricmc.loader.api.FabricLoader;
+import net.fabricmc.loader.api.ModContainer;
 import net.fabricmc.loader.api.metadata.CustomValue;
 import net.fabricmc.loader.api.metadata.ModMetadata;
 import net.minecraft.component.ComponentType;
@@ -22,16 +25,26 @@ import net.minecraft.registry.Registry;
 import net.minecraft.util.Identifier;
 
 import java.lang.invoke.MethodHandles;
+import java.util.Optional;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 
 public class BazaarUtils implements ClientModInitializer {
-    public static IEventBus EVENT_BUS = new EventBus();
-    public static final String MODID = "bazaarutils";
+    public static final String MOD_ID = "bazaarutils";
     public static final String MOD_NAME = "Bazaar Utils";
+
+    public static ModContainer SELF = FabricLoader.getInstance().getModContainer(MOD_ID).get();
+
+    public static Configurator CONFIGURATOR = new Configurator(MOD_ID);
+
+    public static ResourcefulConfig config = BUConfig.register(CONFIGURATOR);
+
     public static boolean updatedMajorVersion = false;
+
     @Getter
     private static String updateNotes;
+
+    public static IEventBus EVENT_BUS = new EventBus();
     public static ScheduledExecutorService BUExecutorService = Executors.newSingleThreadScheduledExecutor();
 
     public static ComponentType<String> CUSTOM_SIZE_COMPONENT;
@@ -41,8 +54,6 @@ public class BazaarUtils implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         registerDataComponents();
-
-        BUConfig.register();
 
         BUCompatibilityHelper.initializePatches();
 
@@ -56,12 +67,12 @@ public class BazaarUtils implements ClientModInitializer {
     private static void registerDataComponents() {
         CUSTOM_SIZE_COMPONENT = Registry.register(
                 Registries.DATA_COMPONENT_TYPE,
-                Identifier.of(BazaarUtils.MODID, "custom_size"),
+                Identifier.of(BazaarUtils.MOD_ID, "custom_size"),
                 ComponentType.<String>builder().codec(Codec.STRING).build()
         );
         CUSTOM_SHOWPRICECHART_COMPONENT = Registry.register(
                 Registries.DATA_COMPONENT_TYPE,
-                Identifier.of(BazaarUtils.MODID, "has_price_chart"),
+                Identifier.of(BazaarUtils.MOD_ID, "has_price_chart"),
                 ComponentType.<Boolean>builder().codec(Codec.BOOL).build()
         );
     }
@@ -89,7 +100,7 @@ public class BazaarUtils implements ClientModInitializer {
 
     }
     private void getModProperties(){
-        FabricLoader.getInstance().getModContainer(MODID).ifPresent(modContainer -> {
+        FabricLoader.getInstance().getModContainer(MOD_ID).ifPresent(modContainer -> {
             ModMetadata metadata = modContainer.getMetadata();
 
             CustomValue updateNotesValue = metadata.getCustomValue("latestMajorUpdateNotes");
