+++ src/main/java/com/github/mkram17/bazaarutils/features/Bookmark.java
@@ -53,8 +53,8 @@ public class Bookmark extends CustomItemButton {
 
     private static final int SIGN_SLOT_NUMBER = 45;
 
-    private static final Identifier BASE = Identifier.tryParse(BazaarUtils.MODID, "widget/bookmark_widget_base");
-    private static final Identifier HOVER = Identifier.tryParse(BazaarUtils.MODID, "widget/bookmark_widget_hover");
+    private static final Identifier BASE = Identifier.tryParse(BazaarUtils.MOD_ID, "widget/bookmark_widget_base");
+    private static final Identifier HOVER = Identifier.tryParse(BazaarUtils.MOD_ID, "widget/bookmark_widget_hover");
 
     public static final ButtonTextures SLOT_BUTTON_TEXTURES = new ButtonTextures(
             BASE,
